+++ resources/views/project.blade.php
@@ -52,7 +52,7 @@
                                     </svg>&emsp;Edit
                                 </button>
                             </a>
-                            <a href="/delete-project/{{ $pro['id'] }}">
+                            <a href="/delete-project/{{ $pro['id'] }}" onclick = "return confirm('Are you sure you want to delete the data?\n\nThis action cannot be undone!');">
                                 <button type="button" class="btn btn-outline-danger">
                                     <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor"
                                         class="bi bi-trash" viewBox="0 0 16 16">
